Add ViewModel (#11)
* Fix lint issues

* Fix lint issues

* Pass object from freaks to details

* Fix more lint issues

* Add ViewModel

* ModelView PR changes

* ModelView PR changes - features gradle

* Resolve PR#10 conflicts
