+++ features/freaks/build.gradle.kts
@@ -8,3 +8,12 @@ plugins {
 android {
     addProductFlavours(this)
 }
+object PluginsVersions {
+    const val LIFECYCLE_VIEW_MODEL = "2.3.1"
+}
+
+
+dependencies {
+    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1")
+    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:${PluginsVersions.LIFECYCLE_VIEW_MODEL}")
+}
