+++ features/freaks/src/main/res/layout/fragment_freak_details.xml
@@ -2,6 +2,8 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android">
 
     <TextView
+        android:id="@+id/details_text_view"
+        android:textIsSelectable="false"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
